example_project/pyproject.toml
[tool.poetry]
name = "example_project"
version = "0.1.0"
description = ""
authors = ["Sai Sasidhar Maddali"]
readme = "readme.md"

[tool.poetry.dependencies]
python = "^3.7"
numpy = "1.18"
pyzmq = "22.2.1"

[tool.poetry.dev-dependencies]
black = "*"
isort = "*"

[tool.poetry.scripts]
main = "main:main"
retval = "main:retval"

[tool.poetry-plugin-appimage]
miniconda = "4.8.3"
python = "3.7"
categories = "ConsoleOnly;"
dependency-group = "main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
